pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=2.0.0,<3.0.0"]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.3.0"

[project]
authors = [
    {email = "kirillbogomolov.ric@yandex.ru", name = "Kirill Bogomolov"},
]
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "django-split-settings (>=1.3.2,<2.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "ipython (>=9.5.0,<10.0.0)",
    "django-debug-toolbar (>=6.0.0,<7.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "faker (>=37.6.0,<38.0.0)",
]
description = ""
name = "test-pronin-team"
readme = "README.md"
requires-python = ">=3.11,<4.0"
version = "0.1.0"
